Split test_trigger_embargo.py into per-operation files
- Remove 1028-line monolith test_trigger_embargo.py (57 tests)
- Add test_trigger_propose_embargo.py (15 tests)
- Add test_trigger_propose_embargo_revision.py (8 tests)
- Add test_trigger_accept_embargo.py (13 tests)
- Add test_trigger_reject_embargo.py (8 tests)
- Add test_trigger_terminate_embargo.py (13 tests)
- Move shared fixtures (client_triggers, case_without_participant,
  case_with_embargo, case_with_proposal, _add_case_manager helper)
  to routers/conftest.py
- Inline _no_outbox_delivery autouse fixture per file to avoid
  applying the outbox patch across the entire routers directory
- Absorb TestTriggerEmbargoOutboxScheduling tests into their
  respective operation files as standalone functions

Closes #493
