+++ test/adapters/driving/fastapi/routers/conftest.py
@@ -20,12 +20,27 @@ from vultron.adapters.driven.datalayer_sqlite import (
     SqliteDataLayer,
     reset_datalayer,
 )
+from vultron.adapters.driven.trigger_activity_adapter import (
+    TriggerActivityAdapter,
+)
+from vultron.adapters.driving.fastapi.deps import (
+    get_canonical_actor_dl,
+    get_trigger_dl,
+    get_trigger_service,
+)
 from vultron.adapters.driving.fastapi.routers import actors as actors_router
 from vultron.adapters.driving.fastapi.routers import (
     datalayer as datalayer_router,
 )
+from vultron.adapters.driving.fastapi.routers import (
+    trigger_embargo as trigger_embargo_router,
+)
 from vultron.core.models.actor import CoreActor
+from vultron.core.states.em import EM
+from vultron.core.states.roles import CVDRole
+from vultron.core.use_cases.triggers.service import TriggerService
 from vultron.adapters.utils import make_id
+from vultron.wire.as2.factories import em_propose_embargo_activity
 from vultron.wire.as2.vocab.base.objects.activities.transitive import as_Offer
 from vultron.wire.as2.vocab.base.objects.actors import (
     as_Organization,
@@ -34,6 +49,9 @@ from vultron.wire.as2.vocab.base.objects.actors import (
     as_Application,
     as_Group,
 )
+from vultron.wire.as2.vocab.objects.case_participant import CaseParticipant
+from vultron.wire.as2.vocab.objects.embargo_event import EmbargoEvent
+from vultron.wire.as2.vocab.objects.vulnerability_case import VulnerabilityCase
 from vultron.wire.as2.vocab.objects.vulnerability_report import (
     VulnerabilityReport,
 )
@@ -145,3 +163,80 @@ def report():
 def offer(report):
     # use current parameter name `object_` (legacy name was `as_object` / JSON `object`)
     return as_Offer(actor="urn:uuid:test-actor", object_=report)
+
+
+# ---------------------------------------------------------------------------
+# Shared embargo test helpers and fixtures
+# ---------------------------------------------------------------------------
+
+
+def _add_case_manager(case: VulnerabilityCase, dl) -> as_Service:
+    """Add a CASE_MANAGER participant to *case* and return the case actor."""
+    case_actor = as_Service(name=f"Case Actor for {case.name}")
+    dl.create(case_actor)
+    cm_participant = CaseParticipant(
+        attributed_to=case_actor.id_,
+        context=case.id_,
+        case_roles=[CVDRole.CASE_MANAGER],
+    )
+    dl.create(cm_participant)
+    case.actor_participant_index[case_actor.id_] = cm_participant.id_
+    dl.save(case)
+    return case_actor
+
+
+@pytest.fixture
+def client_triggers(dl):
+    """TestClient wired to the trigger_embargo router with overridden deps."""
+    app = FastAPI()
+    app.include_router(trigger_embargo_router.router)
+    app.dependency_overrides[get_trigger_service] = lambda: TriggerService(
+        dl, trigger_activity=TriggerActivityAdapter(dl)
+    )
+    app.dependency_overrides[get_trigger_dl] = lambda: dl
+    app.dependency_overrides[get_canonical_actor_dl] = lambda: dl
+    client = TestClient(app)
+    yield client
+    app.dependency_overrides = {}
+
+
+@pytest.fixture
+def case_without_participant(dl):
+    """A VulnerabilityCase with a Case Manager but no participant for the actor."""
+    case_obj = VulnerabilityCase(name="TEST-CASE-NO-PARTICIPANT")
+    dl.create(case_obj)
+    _add_case_manager(case_obj, dl)
+    return case_obj
+
+
+@pytest.fixture
+def case_with_embargo(dl, actor):
+    """A VulnerabilityCase with an active EmbargoEvent."""
+    case_obj = VulnerabilityCase(name="EMBARGO-CASE-001")
+    embargo = EmbargoEvent(context=case_obj.id_)
+    dl.create(embargo)
+    case_obj.set_embargo(embargo.id_)
+    case_obj.current_status.em_state = EM.ACTIVE
+    dl.create(case_obj)
+    _add_case_manager(case_obj, dl)
+    return case_obj, embargo
+
+
+@pytest.fixture
+def case_with_proposal(dl, actor):
+    """A VulnerabilityCase with a pending EmProposeEmbargoActivity in EM.PROPOSED state."""
+    case_obj = VulnerabilityCase(
+        name="PROPOSAL-CASE-001",
+        attributed_to=actor.id_,
+    )
+    embargo = EmbargoEvent(context=case_obj.id_)
+    dl.create(embargo)
+    proposal = em_propose_embargo_activity(
+        embargo, context=case_obj.id_, actor=actor.id_
+    )
+    dl.create(proposal)
+    case_obj.current_status.em_state = EM.PROPOSED
+    case_obj.proposed_embargoes.append(embargo.id_)
+    dl.create(case_obj)
+    _add_case_manager(case_obj, dl)
+    return case_obj, proposal, embargo
