--- test/adapters/driving/fastapi/routers/test_trigger_embargo.py
@@ -1,1013 +1,0 @@
-#!/usr/bin/env python
-
-#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
-#  - see Contributors.md for a full list of Contributors
-#  - see ContributionInstructions.md for information on how you can Contribute to this project
-#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
-#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
-#  with this Software or contact permission@sei.cmu.edu for full terms.
-#  Created, in part, with funding and support from the United States Government
-#  (see Acknowledgments file). This program may include and/or can make use of
-#  certain third party source code, object code, documentation and other files
-#  ("Third Party Software"). See LICENSE.md for more details.
-#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
-#  U.S. Patent and Trademark Office by Carnegie Mellon University
-
-"""
-Tests for the embargo trigger endpoints
-(POST /actors/{actor_id}/trigger/{propose,accept,reject,terminate,propose-revision}-embargo).
-
-Verifies TB-01 through TB-07 requirements from specs/triggerable-behaviors.yaml.
-"""
-
-import pytest
-from fastapi import FastAPI, status
-from fastapi.testclient import TestClient
-from unittest.mock import AsyncMock, patch
-
-from vultron.adapters.driven.db_record import object_to_record
-from vultron.adapters.driving.fastapi.deps import (
-    get_canonical_actor_dl,
-    get_trigger_dl,
-    get_trigger_service,
-)
-from vultron.adapters.driving.fastapi.routers import (
-    trigger_embargo as trigger_embargo_router,
-)
-from vultron.core.use_cases.triggers.service import TriggerService
-from vultron.adapters.driven.trigger_activity_adapter import (
-    TriggerActivityAdapter,
-)
-from vultron.wire.as2.factories import em_propose_embargo_activity
-from vultron.wire.as2.vocab.base.objects.actors import as_Service
-from vultron.wire.as2.vocab.objects.case_participant import CaseParticipant
-from vultron.wire.as2.vocab.objects.embargo_event import EmbargoEvent
-from vultron.wire.as2.vocab.objects.vulnerability_case import VulnerabilityCase
-from vultron.core.states.em import EM
-from vultron.core.states.roles import CVDRole
-
-FUTURE_END_TIME = "2099-12-01T00:00:00Z"
-
-
-def _add_case_manager(case: VulnerabilityCase, dl) -> as_Service:
-    """Add a CASE_MANAGER participant to *case* and return the case actor."""
-    case_actor = as_Service(name=f"Case Actor for {case.name}")
-    dl.create(case_actor)
-    cm_participant = CaseParticipant(
-        attributed_to=case_actor.id_,
-        context=case.id_,
-        case_roles=[CVDRole.CASE_MANAGER],
-    )
-    dl.create(cm_participant)
-    case.actor_participant_index[case_actor.id_] = cm_participant.id_
-    dl.save(case)
-    return case_actor
-
-
-# ---------------------------------------------------------------------------
-# Module-level fixture: suppress outbox delivery retries
-# ---------------------------------------------------------------------------
-
-
-@pytest.fixture(autouse=True)
-def _no_outbox_delivery():
-    """Suppress real outbox delivery for every test in this module.
-
-    ``outbox_handler`` uses HTTP with exponential-backoff retries. When
-    tests run with non-existent recipient URLs the retry sleeps add ~3.5 s
-    per test. Patching to a no-op ``AsyncMock`` eliminates that overhead
-    while keeping the scheduler logic testable.
-
-    Tests in ``TestEmbargoTriggerOutboxScheduling`` that need a trackable
-    mock use ``unittest.mock.patch`` as a context manager inside the test
-    body, which overrides this fixture's patch for the duration of that
-    context.
-    """
-    with patch(
-        "vultron.adapters.driving.fastapi.routers"
-        ".trigger_embargo.outbox_handler",
-        new_callable=AsyncMock,
-    ):
-        yield
-
-
-# ---------------------------------------------------------------------------
-# Fixtures
-# ---------------------------------------------------------------------------
-
-
-@pytest.fixture
-def client_triggers(dl):
-    app = FastAPI()
-    app.include_router(trigger_embargo_router.router)
-    app.dependency_overrides[get_trigger_service] = lambda: TriggerService(
-        dl, trigger_activity=TriggerActivityAdapter(dl)
-    )
-    app.dependency_overrides[get_trigger_dl] = lambda: dl
-    app.dependency_overrides[get_canonical_actor_dl] = lambda: dl
-    client = TestClient(app)
-    yield client
-    app.dependency_overrides = {}
-
-
-@pytest.fixture
-def case_without_participant(dl):
-    """A VulnerabilityCase with a Case Manager but no participant for the actor."""
-    case_obj = VulnerabilityCase(name="TEST-CASE-NO-PARTICIPANT")
-    dl.create(case_obj)
-    _add_case_manager(case_obj, dl)
-    return case_obj
-
-
-@pytest.fixture
-def case_with_embargo(dl, actor):
-    """A VulnerabilityCase with an active EmbargoEvent."""
-    case_obj = VulnerabilityCase(name="EMBARGO-CASE-001")
-    embargo = EmbargoEvent(context=case_obj.id_)
-    dl.create(embargo)
-    case_obj.set_embargo(embargo.id_)
-    case_obj.current_status.em_state = EM.ACTIVE
-    dl.create(case_obj)
-    _add_case_manager(case_obj, dl)
-    return case_obj, embargo
-
-
-@pytest.fixture
-def case_with_proposal(dl, actor):
-    """A VulnerabilityCase with a pending EmProposeEmbargoActivity in EM.PROPOSED state."""
-    case_obj = VulnerabilityCase(
-        name="PROPOSAL-CASE-001",
-        attributed_to=actor.id_,
-    )
-    embargo = EmbargoEvent(context=case_obj.id_)
-    dl.create(embargo)
-    proposal = em_propose_embargo_activity(
-        embargo, context=case_obj.id_, actor=actor.id_
-    )
-    dl.create(proposal)
-    case_obj.current_status.em_state = EM.PROPOSED
-    case_obj.proposed_embargoes.append(embargo.id_)
-    dl.create(case_obj)
-    _add_case_manager(case_obj, dl)
-    return case_obj, proposal, embargo
-
-
-# ===========================================================================
-# Tests for trigger/propose-embargo
-# ===========================================================================
-
-
-def test_trigger_propose_embargo_returns_202(
-    client_triggers, actor, case_without_participant
-):
-    """TB-01-002: POST /actors/{id}/trigger/propose-embargo returns HTTP 202."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={
-            "case_id": case_without_participant.id_,
-            "end_time": FUTURE_END_TIME,
-        },
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-
-def test_trigger_propose_embargo_response_contains_activity_key(
-    client_triggers, actor, case_without_participant
-):
-    """TB-04-001: Successful trigger response body contains 'activity' key."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={
-            "case_id": case_without_participant.id_,
-            "end_time": FUTURE_END_TIME,
-        },
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-    data = resp.json()
-    assert "activity" in data
-    assert data["activity"] is not None
-
-
-def test_trigger_propose_embargo_missing_case_id_returns_422(
-    client_triggers, actor
-):
-    """TB-03-001: Request missing case_id returns HTTP 422."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={"end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_propose_embargo_missing_end_time_returns_422(
-    client_triggers, actor, case_without_participant
-):
-    """propose-embargo without end_time returns HTTP 422."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={"case_id": case_without_participant.id_},
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_propose_embargo_naive_end_time_returns_422(
-    client_triggers, actor, case_without_participant
-):
-    """propose-embargo with timezone-naive end_time returns HTTP 422."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={
-            "case_id": case_without_participant.id_,
-            "end_time": "2099-12-01T00:00:00",
-        },
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_propose_embargo_past_end_time_returns_422(
-    client_triggers, actor, case_without_participant
-):
-    """propose-embargo with a past end_time returns HTTP 422."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={
-            "case_id": case_without_participant.id_,
-            "end_time": "2020-01-01T00:00:00Z",
-        },
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_propose_embargo_ignores_unknown_fields(
-    client_triggers, actor, case_without_participant
-):
-    """TB-03-002: Unknown fields in request body are silently ignored."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={
-            "case_id": case_without_participant.id_,
-            "end_time": FUTURE_END_TIME,
-            "unknown_xyz": 99,
-        },
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-
-def test_trigger_propose_embargo_unknown_actor_returns_404(client_triggers):
-    """TB-01-003: Unknown actor_id returns HTTP 404 with structured body."""
-    resp = client_triggers.post(
-        "/actors/nonexistent-actor/trigger/propose-embargo",
-        json={"case_id": "urn:uuid:any-case", "end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-    data = resp.json()
-    assert data["detail"]["error"] == "NotFound"
-
-
-def test_trigger_propose_embargo_unknown_case_returns_404(
-    client_triggers, actor
-):
-    """TB-01-003: Unknown case_id returns HTTP 404."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={
-            "case_id": "urn:uuid:nonexistent-case",
-            "end_time": FUTURE_END_TIME,
-        },
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-
-
-def test_trigger_propose_embargo_invalid_case_id_returns_422(
-    client_triggers, actor
-):
-    """propose-embargo with a non-URI case_id returns HTTP 422."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={"case_id": "not-a-uri", "end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_propose_embargo_adds_activity_to_outbox(
-    client_triggers, dl, actor, case_without_participant
-):
-    """TB-07-001: Successful trigger adds a new activity to actor's outbox."""
-    outbox_before = set(dl.outbox_list())
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={
-            "case_id": case_without_participant.id_,
-            "end_time": FUTURE_END_TIME,
-        },
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    outbox_after = set(dl.outbox_list())
-    assert len(outbox_after - outbox_before) >= 1
-
-
-def test_trigger_propose_embargo_updates_em_state_to_proposed(
-    client_triggers, dl, actor, case_without_participant
-):
-    """propose-embargo transitions case EM state from N to P."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={
-            "case_id": case_without_participant.id_,
-            "end_time": FUTURE_END_TIME,
-        },
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    updated_case = dl.read(case_without_participant.id_)
-    assert updated_case.current_status.em_state == EM.PROPOSED
-
-
-def test_trigger_propose_embargo_from_active_updates_em_state_to_revise(
-    client_triggers, dl, actor, case_with_embargo
-):
-    """propose-embargo transitions case EM state from A to R when embargo is active."""
-    case_obj, _ = case_with_embargo
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={"case_id": case_obj.id_, "end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    updated_case = dl.read(case_obj.id_)
-    assert updated_case.current_status.em_state == EM.REVISE
-
-
-def test_trigger_propose_embargo_exited_returns_409(
-    client_triggers, dl, actor, case_without_participant
-):
-    """propose-embargo returns HTTP 409 when EM state is EXITED."""
-    case_obj = dl.read(case_without_participant.id_)
-    case_obj.current_status.em_state = EM.EXITED
-    dl.update(case_obj.id_, object_to_record(case_obj))
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo",
-        json={
-            "case_id": case_without_participant.id_,
-            "end_time": FUTURE_END_TIME,
-        },
-    )
-    assert resp.status_code == status.HTTP_409_CONFLICT
-    data = resp.json()
-    assert data["detail"]["error"] == "Conflict"
-
-
-# ===========================================================================
-# Tests for trigger/accept-embargo
-# ===========================================================================
-
-
-def test_trigger_accept_embargo_returns_202(
-    client_triggers, actor, case_with_proposal
-):
-    """TB-01-002: POST /actors/{id}/trigger/accept-embargo returns HTTP 202."""
-    case_obj, proposal, _ = case_with_proposal
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={"case_id": case_obj.id_, "proposal_id": proposal.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-
-def test_trigger_accept_embargo_response_contains_activity_key(
-    client_triggers, actor, case_with_proposal
-):
-    """TB-04-001: Successful trigger response body contains 'activity' key."""
-    case_obj, proposal, _ = case_with_proposal
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={"case_id": case_obj.id_, "proposal_id": proposal.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-    data = resp.json()
-    assert "activity" in data
-    assert data["activity"] is not None
-
-
-def test_trigger_accept_embargo_object_is_proposal(
-    client_triggers, actor, case_with_proposal
-):
-    """DR-05: Accept activity object_ must be the original proposal, not the embargo event."""
-    case_obj, proposal, _ = case_with_proposal
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={"case_id": case_obj.id_, "proposal_id": proposal.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-    data = resp.json()
-    assert data["activity"]["object"]["id"] == proposal.id_
-
-
-def test_trigger_accept_embargo_missing_case_id_returns_422(
-    client_triggers, actor
-):
-    """TB-03-001: Request missing case_id returns HTTP 422."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={},
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_accept_embargo_ignores_unknown_fields(
-    client_triggers, actor, case_with_proposal
-):
-    """TB-03-002: Unknown fields in request body are silently ignored."""
-    case_obj, proposal, _ = case_with_proposal
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={
-            "case_id": case_obj.id_,
-            "proposal_id": proposal.id_,
-            "extra": "ignored",
-        },
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-
-def test_trigger_accept_embargo_unknown_actor_returns_404(client_triggers):
-    """TB-01-003: Unknown actor_id returns HTTP 404 with structured body."""
-    resp = client_triggers.post(
-        "/actors/nonexistent-actor/trigger/accept-embargo",
-        json={"case_id": "urn:uuid:any-case"},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-    data = resp.json()
-    assert data["detail"]["error"] == "NotFound"
-
-
-def test_trigger_accept_embargo_unknown_case_returns_404(
-    client_triggers, actor
-):
-    """TB-01-003: Unknown case_id returns HTTP 404."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={"case_id": "urn:uuid:nonexistent-case"},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-
-
-def test_trigger_accept_embargo_unknown_proposal_returns_404(
-    client_triggers, actor, case_without_participant
-):
-    """TB-01-003: Unknown proposal_id returns HTTP 404."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={
-            "case_id": case_without_participant.id_,
-            "proposal_id": "urn:uuid:nonexistent-proposal",
-        },
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-
-
-def test_trigger_accept_embargo_no_proposal_returns_404(
-    client_triggers, actor, case_without_participant
-):
-    """accept-embargo returns HTTP 404 when no proposal is found for the case."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={"case_id": case_without_participant.id_},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-
-
-def test_trigger_accept_embargo_adds_activity_to_outbox(
-    client_triggers, dl, actor, case_with_proposal
-):
-    """TB-07-001: Successful trigger adds a new activity to actor's outbox."""
-    case_obj, proposal, _ = case_with_proposal
-    outbox_before = set(dl.outbox_list())
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={"case_id": case_obj.id_, "proposal_id": proposal.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    outbox_after = set(dl.outbox_list())
-    assert len(outbox_after - outbox_before) >= 1
-
-
-def test_trigger_accept_embargo_activates_embargo(
-    client_triggers, dl, actor, case_with_proposal
-):
-    """accept-embargo activates the embargo and sets EM state to ACTIVE."""
-    case_obj, proposal, embargo = case_with_proposal
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={"case_id": case_obj.id_, "proposal_id": proposal.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    updated_case = dl.read(case_obj.id_)
-    assert updated_case.current_status.em_state == EM.ACTIVE
-    assert updated_case.active_embargo is not None
-
-
-def test_trigger_accept_embargo_without_proposal_id_uses_first_proposal(
-    client_triggers, dl, actor, case_with_proposal
-):
-    """accept-embargo without proposal_id finds the first pending proposal."""
-    case_obj, _, _ = case_with_proposal
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/accept-embargo",
-        json={"case_id": case_obj.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    updated_case = dl.read(case_obj.id_)
-    assert updated_case.current_status.em_state == EM.ACTIVE
-
-
-# ===========================================================================
-# Tests for trigger/reject-embargo
-# ===========================================================================
-
-
-def test_trigger_reject_embargo_returns_202(
-    client_triggers, actor, case_with_proposal
-):
-    """TB-01-002: POST /actors/{id}/trigger/reject-embargo returns HTTP 202."""
-    case_obj, proposal, _ = case_with_proposal
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/reject-embargo",
-        json={"case_id": case_obj.id_, "proposal_id": proposal.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-
-def test_trigger_reject_embargo_response_contains_activity_key(
-    client_triggers, actor, case_with_proposal
-):
-    """TB-04-001: Successful trigger response body contains 'activity' key."""
-    case_obj, proposal, _ = case_with_proposal
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/reject-embargo",
-        json={"case_id": case_obj.id_, "proposal_id": proposal.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-    data = resp.json()
-    assert "activity" in data
-    assert data["activity"] is not None
-
-
-def test_trigger_reject_embargo_missing_case_id_returns_422(
-    client_triggers, actor
-):
-    """TB-03-001: Request missing case_id returns HTTP 422."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/reject-embargo",
-        json={},
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_reject_embargo_unknown_actor_returns_404(client_triggers):
-    """TB-01-003: Unknown actor_id returns HTTP 404 with structured body."""
-    resp = client_triggers.post(
-        "/actors/nonexistent-actor/trigger/reject-embargo",
-        json={"case_id": "urn:uuid:any-case"},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-    data = resp.json()
-    assert data["detail"]["error"] == "NotFound"
-
-
-def test_trigger_reject_embargo_unknown_case_returns_404(
-    client_triggers, actor
-):
-    """TB-01-003: Unknown case_id returns HTTP 404."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/reject-embargo",
-        json={"case_id": "urn:uuid:nonexistent-case"},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-
-
-def test_trigger_reject_embargo_no_proposal_returns_404(
-    client_triggers, actor, case_without_participant
-):
-    """reject-embargo returns HTTP 404 when no proposal is found for the case."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/reject-embargo",
-        json={"case_id": case_without_participant.id_},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-
-
-def test_trigger_reject_embargo_sets_em_state_to_no_embargo(
-    client_triggers, dl, actor, case_with_proposal
-):
-    """reject-embargo transitions case EM state from PROPOSED to NO_EMBARGO."""
-    case_obj, proposal, _ = case_with_proposal
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/reject-embargo",
-        json={"case_id": case_obj.id_, "proposal_id": proposal.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    updated_case = dl.read(case_obj.id_)
-    assert updated_case.current_status.em_state == EM.NO_EMBARGO
-
-
-def test_trigger_reject_embargo_adds_activity_to_outbox(
-    client_triggers, dl, actor, case_with_proposal
-):
-    """TB-07-001: Successful trigger adds a new activity to actor's outbox."""
-    case_obj, proposal, _ = case_with_proposal
-    outbox_before = set(dl.outbox_list())
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/reject-embargo",
-        json={"case_id": case_obj.id_, "proposal_id": proposal.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    outbox_after = set(dl.outbox_list())
-    assert len(outbox_after - outbox_before) >= 1
-
-
-# ===========================================================================
-# Tests for trigger/propose-embargo-revision
-# ===========================================================================
-
-
-def test_trigger_propose_embargo_revision_returns_202(
-    client_triggers, actor, case_with_embargo
-):
-    """TB-01-002: POST /actors/{id}/trigger/propose-embargo-revision returns 202."""
-    case_obj, _ = case_with_embargo
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo-revision",
-        json={"case_id": case_obj.id_, "end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-
-def test_trigger_propose_embargo_revision_response_contains_activity_key(
-    client_triggers, actor, case_with_embargo
-):
-    """TB-04-001: Successful trigger response body contains 'activity' key."""
-    case_obj, _ = case_with_embargo
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo-revision",
-        json={"case_id": case_obj.id_, "end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-    data = resp.json()
-    assert "activity" in data
-    assert data["activity"] is not None
-
-
-def test_trigger_propose_embargo_revision_missing_case_id_returns_422(
-    client_triggers, actor
-):
-    """TB-03-001: Request missing case_id returns HTTP 422."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo-revision",
-        json={"end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_propose_embargo_revision_missing_end_time_returns_422(
-    client_triggers, actor, case_with_embargo
-):
-    """TB-03-001: Request missing end_time returns HTTP 422."""
-    case_obj, _ = case_with_embargo
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo-revision",
-        json={"case_id": case_obj.id_},
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_propose_embargo_revision_unknown_actor_returns_404(
-    client_triggers,
-):
-    """TB-01-003: Unknown actor_id returns HTTP 404 with structured body."""
-    resp = client_triggers.post(
-        "/actors/nonexistent-actor/trigger/propose-embargo-revision",
-        json={"case_id": "urn:uuid:any-case", "end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-    data = resp.json()
-    assert data["detail"]["error"] == "NotFound"
-
-
-def test_trigger_propose_embargo_revision_unknown_case_returns_404(
-    client_triggers, actor
-):
-    """TB-01-003: Unknown case_id returns HTTP 404."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo-revision",
-        json={
-            "case_id": "urn:uuid:nonexistent-case",
-            "end_time": FUTURE_END_TIME,
-        },
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-
-
-def test_trigger_propose_embargo_revision_non_active_em_state_returns_409(
-    client_triggers, actor, case_without_participant
-):
-    """propose-embargo-revision returns 409 when case EM state is not ACTIVE/REVISE."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo-revision",
-        json={
-            "case_id": case_without_participant.id_,
-            "end_time": FUTURE_END_TIME,
-        },
-    )
-    assert resp.status_code == status.HTTP_409_CONFLICT
-    data = resp.json()
-    assert data["detail"]["error"] == "Conflict"
-
-
-def test_trigger_propose_embargo_revision_sets_em_state_to_revise(
-    client_triggers, dl, actor, case_with_embargo
-):
-    """propose-embargo-revision transitions case EM state from ACTIVE to REVISE."""
-    case_obj, _ = case_with_embargo
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo-revision",
-        json={"case_id": case_obj.id_, "end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    updated_case = dl.read(case_obj.id_)
-    assert updated_case.current_status.em_state == EM.REVISE
-
-
-def test_trigger_propose_embargo_revision_adds_activity_to_outbox(
-    client_triggers, dl, actor, case_with_embargo
-):
-    """TB-07-001: Successful trigger adds a new activity to actor's outbox."""
-    case_obj, _ = case_with_embargo
-    outbox_before = set(dl.outbox_list())
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/propose-embargo-revision",
-        json={"case_id": case_obj.id_, "end_time": FUTURE_END_TIME},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    outbox_after = set(dl.outbox_list())
-    assert len(outbox_after - outbox_before) >= 1
-
-
-# ===========================================================================
-# Tests for trigger/terminate-embargo
-# ===========================================================================
-
-
-def test_trigger_terminate_embargo_returns_202(
-    client_triggers, actor, case_with_embargo
-):
-    """TB-01-002: POST /actors/{id}/trigger/terminate-embargo returns HTTP 202."""
-    case_obj, _ = case_with_embargo
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={"case_id": case_obj.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-
-def test_trigger_terminate_embargo_response_contains_activity_key(
-    client_triggers, actor, case_with_embargo
-):
-    """TB-04-001: Successful trigger response body contains 'activity' key."""
-    case_obj, _ = case_with_embargo
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={"case_id": case_obj.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-    data = resp.json()
-    assert "activity" in data
-    assert data["activity"] is not None
-
-
-def test_trigger_terminate_embargo_missing_case_id_returns_422(
-    client_triggers, actor
-):
-    """TB-03-001: Request missing case_id returns HTTP 422."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={},
-    )
-    assert resp.status_code == status.HTTP_422_UNPROCESSABLE_CONTENT
-
-
-def test_trigger_terminate_embargo_ignores_unknown_fields(
-    client_triggers, actor, case_with_embargo
-):
-    """TB-03-002: Unknown fields in request body are silently ignored."""
-    case_obj, _ = case_with_embargo
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={"case_id": case_obj.id_, "extra": "ignored"},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-
-def test_trigger_terminate_embargo_unknown_actor_returns_404(client_triggers):
-    """TB-01-003: Unknown actor_id returns HTTP 404 with structured body."""
-    resp = client_triggers.post(
-        "/actors/nonexistent-actor/trigger/terminate-embargo",
-        json={"case_id": "urn:uuid:any-case"},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-    data = resp.json()
-    assert data["detail"]["error"] == "NotFound"
-
-
-def test_trigger_terminate_embargo_unknown_case_returns_404(
-    client_triggers, actor
-):
-    """TB-01-003: Unknown case_id returns HTTP 404."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={"case_id": "urn:uuid:nonexistent-case"},
-    )
-    assert resp.status_code == status.HTTP_404_NOT_FOUND
-
-
-def test_trigger_terminate_embargo_no_active_embargo_returns_409(
-    client_triggers, actor, case_without_participant
-):
-    """terminate-embargo returns HTTP 409 when case has no active embargo."""
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={"case_id": case_without_participant.id_},
-    )
-    assert resp.status_code == status.HTTP_409_CONFLICT
-    data = resp.json()
-    assert data["detail"]["error"] == "Conflict"
-
-
-def test_trigger_terminate_embargo_adds_activity_to_outbox(
-    client_triggers, dl, actor, case_with_embargo
-):
-    """TB-07-001: Successful trigger adds a new activity to actor's outbox."""
-    case_obj, _ = case_with_embargo
-    outbox_before = set(dl.outbox_list())
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={"case_id": case_obj.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    outbox_after = set(dl.outbox_list())
-    assert len(outbox_after - outbox_before) >= 1
-
-
-def test_trigger_terminate_embargo_updates_em_state_to_exited(
-    client_triggers, dl, actor, case_with_embargo
-):
-    """terminate-embargo transitions case EM state to EXITED."""
-    case_obj, _ = case_with_embargo
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={"case_id": case_obj.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    updated_case = dl.read(case_obj.id_)
-    assert updated_case.current_status.em_state == EM.EXITED
-
-
-def test_trigger_terminate_embargo_clears_active_embargo(
-    client_triggers, dl, actor, case_with_embargo
-):
-    """terminate-embargo clears the active_embargo field on the case."""
-    case_obj, _ = case_with_embargo
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={"case_id": case_obj.id_},
-    )
-    assert resp.status_code == status.HTTP_202_ACCEPTED
-
-    updated_case = dl.read(case_obj.id_)
-    assert updated_case.active_embargo is None
-
-
-def test_trigger_terminate_embargo_invalid_em_state_returns_409(
-    client_triggers, dl, actor, case_with_embargo
-):
-    """terminate-embargo returns HTTP 409 when EM state is not ACTIVE or REVISE.
-
-    Guards against bypassing the EM machine: even if active_embargo is set,
-    terminating from a state with no TERMINATE transition (e.g. PROPOSED) must
-    be rejected.
-    """
-    case_obj, _ = case_with_embargo
-    stored = dl.read(case_obj.id_)
-    stored.current_status.em_state = EM.PROPOSED
-    dl.update(stored.id_, object_to_record(stored))
-
-    resp = client_triggers.post(
-        f"/actors/{actor.id_}/trigger/terminate-embargo",
-        json={"case_id": case_obj.id_},
-    )
-    assert resp.status_code == status.HTTP_409_CONFLICT
-    data = resp.json()
-    assert data["detail"]["error"] == "Conflict"
-
-
-# ===========================================================================
-# Tests for outbox delivery scheduling (D5-6-TRIGDELIV)
-# ===========================================================================
-
-
-class TestTriggerEmbargoOutboxScheduling:
-    """D5-6-TRIGDELIV: embargo trigger endpoints must schedule outbox_handler."""
-
-    def test_propose_embargo_schedules_outbox_handler(
-        self, client_triggers, dl, actor, case_without_participant
-    ):
-        """propose-embargo schedules outbox delivery after execution."""
-        from unittest.mock import AsyncMock, patch
-
-        with patch(
-            "vultron.adapters.driving.fastapi.routers"
-            ".trigger_embargo.outbox_handler",
-            new_callable=AsyncMock,
-        ) as mock_outbox:
-            resp = client_triggers.post(
-                f"/actors/{actor.id_}/trigger/propose-embargo",
-                json={
-                    "case_id": case_without_participant.id_,
-                    "end_time": FUTURE_END_TIME,
-                },
-            )
-        assert resp.status_code == status.HTTP_202_ACCEPTED
-        mock_outbox.assert_called_once()
-        assert mock_outbox.call_args.args[0] == actor.id_
-        assert mock_outbox.call_args.args[1] is dl
-        assert mock_outbox.call_args.args[2] is dl
-
-    def test_accept_embargo_schedules_outbox_handler(
-        self, client_triggers, dl, actor, case_with_proposal
-    ):
-        """accept-embargo schedules outbox delivery after execution."""
-        from unittest.mock import AsyncMock, patch
-
-        case_obj, proposal, _ = case_with_proposal
-        with patch(
-            "vultron.adapters.driving.fastapi.routers"
-            ".trigger_embargo.outbox_handler",
-            new_callable=AsyncMock,
-        ) as mock_outbox:
-            resp = client_triggers.post(
-                f"/actors/{actor.id_}/trigger/accept-embargo",
-                json={
-                    "case_id": case_obj.id_,
-                    "proposal_id": proposal.id_,
-                },
-            )
-        assert resp.status_code == status.HTTP_202_ACCEPTED
-        mock_outbox.assert_called_once()
-        assert mock_outbox.call_args.args[0] == actor.id_
-        assert mock_outbox.call_args.args[1] is dl
-        assert mock_outbox.call_args.args[2] is dl
-
-    def test_terminate_embargo_schedules_outbox_handler(
-        self, client_triggers, dl, actor, case_with_embargo
-    ):
-        """terminate-embargo schedules outbox delivery after execution."""
-        from unittest.mock import AsyncMock, patch
-
-        case_obj, _ = case_with_embargo
-        with patch(
-            "vultron.adapters.driving.fastapi.routers"
-            ".trigger_embargo.outbox_handler",
-            new_callable=AsyncMock,
-        ) as mock_outbox:
-            resp = client_triggers.post(
-                f"/actors/{actor.id_}/trigger/terminate-embargo",
-                json={"case_id": case_obj.id_},
-            )
-        assert resp.status_code == status.HTTP_202_ACCEPTED
-        mock_outbox.assert_called_once()
-        assert mock_outbox.call_args.args[0] == actor.id_
-        assert mock_outbox.call_args.args[1] is dl
-        assert mock_outbox.call_args.args[2] is dl
